Moved Default Entities to DefaultEntities.ts
